🚧 extract grapher and core table types into types package
