+++ packages/@ourworldindata/core-table/src/CoreTable.test.ts
@@ -1,8 +1,7 @@
 #! /usr/bin/env jest
 
 import { CoreTable } from "./CoreTable.js"
-import { TransformType } from "./CoreTableConstants.js"
-import { ColumnTypeNames } from "./CoreColumnDef.js"
+import { TransformType, ColumnTypeNames } from "@ourworldindata/types"
 import { ErrorValueTypes, isNotErrorValue } from "./ErrorValues.js"
 
 const sampleCsv = `country,population
