+++ packages/@ourworldindata/grapher/src/index.ts
@@ -1,6 +1,5 @@
 import "./core/grapher.scss"
 
-export { BinningStrategy } from "./color/BinningStrategy"
 export { binningStrategyLabels } from "./color/BinningStrategies"
 export {
     NumericBin,
@@ -9,7 +8,6 @@ export {
 } from "./color/ColorScaleBin"
 export { ChartDimension } from "./chart/ChartDimension"
 export {
-    ChartTypeName,
     GRAPHER_EMBEDDED_FIGURE_ATTR,
     GRAPHER_EMBEDDED_FIGURE_CONFIG_ATTR,
     GRAPHER_PAGE_BODY_CLASS,
@@ -20,22 +18,9 @@ export {
     STATIC_EXPORT_DETAIL_SPACING,
     DEFAULT_GRAPHER_ENTITY_TYPE,
     CookieKey,
-    EntitySelectionMode,
-    StackMode,
     BASE_FONT_SIZE,
-    FacetStrategy,
-    FacetAxisDomain,
-    SeriesStrategy,
-    MissingDataStrategy,
     ThereWasAProblemLoadingThisChart,
-    type SeriesColorMap,
-    GrapherTabOption,
-    ScaleType,
-    type RelatedQuestionsConfig,
-    type Topic,
     WorldEntityName,
-    ScatterPointLabelStrategy,
-    type SeriesName,
     getVariableDataRoute,
     getVariableMetadataRoute,
     Patterns,
@@ -46,13 +31,11 @@ export {
 export { ColorScale } from "./color/ColorScale"
 export { ColorScaleConfig } from "./color/ColorScaleConfig"
 export { ColorScheme } from "./color/ColorScheme"
-export { ColorSchemeName } from "./color/ColorConstants"
 export {
     getColorNameOwidDistinctAndSemanticPalettes,
     getColorNameOwidDistinctLinesAndSemanticPalettes,
 } from "./color/CustomSchemes"
 export { ColorSchemes } from "./color/ColorSchemes"
-export type { ComparisonLineConfig } from "./scatterCharts/ComparisonLine"
 export { DimensionSlot } from "./chart/DimensionSlot"
 export { EntityPicker } from "./controls/entityPicker/EntityPicker"
 export type { EntityPickerManager } from "./controls/entityPicker/EntityPickerConstants"
@@ -70,13 +53,6 @@ export {
     getErrorMessageRelatedQuestionUrl,
 } from "./core/Grapher"
 export { GrapherAnalytics, EventCategory } from "./core/GrapherAnalytics"
-export {
-    type GrapherInterface,
-    type GrapherQueryParams,
-    type LegacyGrapherInterface,
-    type LegacyGrapherQueryParams,
-    grapherKeysToSerialize,
-} from "./core/GrapherInterface"
 import fuzzysort from "fuzzysort"
 export { fuzzysort }
 export { highlight } from "./controls/FuzzySearch"
@@ -84,11 +60,9 @@ export { hydrateGlobalEntitySelectorIfAny } from "./controls/globalEntitySelecto
 export { legacyToCurrentGrapherUrl } from "./core/GrapherUrlMigrations"
 export { legacyToOwidTableAndDimensions } from "./core/LegacyToOwidTable"
 export { LoadingIndicator } from "./loadingIndicator/LoadingIndicator"
-export { LogoOption } from "./captionedChart/Logos"
 export { MapChart } from "./mapCharts/MapChart"
 export { MapConfig } from "./mapCharts/MapConfig"
 export {
-    MapProjectionName,
     MapProjectionLabels,
     MapProjectionGeos,
 } from "./mapCharts/MapProjections"
