+++ packages/@ourworldindata/types/src/index.ts
@@ -18,6 +18,7 @@ export {
     type OwidVariableId,
     type RawPageview,
     type UserCountryInformation,
+    type QueryParams,
 } from "./domainTypes/Various.js"
 export {
     type DataValueConfiguration,
@@ -78,6 +79,31 @@ export {
     type Time,
     type TimeBound,
     type TimeBounds,
+    type TickFormattingOptions,
+    BinningStrategy,
+    type ColorScaleConfigInterface,
+    type ColorSchemeName,
+    type ChartTypeName,
+    type GrapherTabOption,
+    type StackMode,
+    type EntitySelectionMode,
+    type ScatterPointLabelStrategy,
+    type RelatedQuestionsConfig,
+    type FacetStrategy,
+    type SeriesColorMap,
+    type FacetAxisDomain,
+    type AnnotationFieldsInTitle,
+    type MissingDataStrategy,
+    type SeriesStrategy,
+    type GrapherInterface,
+    grapherKeysToSerialize,
+    type GrapherQueryParams,
+    type LegacyGrapherInterface,
+    MapProjectionName,
+    LogoOption,
+    type ComparisonLineConfig,
+    type AxisConfigInterface,
+    GrapherStaticFormat,
 } from "./grapherTypes/GrapherTypes.js"
 
 export {
@@ -336,3 +362,34 @@ export type {
     OwidVariableDataTableConfigInterface,
     OwidChartDimensionInterface,
 } from "./OwidVariableDisplayConfigInterface.js"
+
+export {
+    type TableSlug,
+    type ColumnSlugs,
+    type TimeTolerance,
+    type CoreRow,
+    InputType,
+    TransformType,
+    JsTypes,
+    type CsvString,
+    type CoreValueType,
+    type CoreColumnStore,
+    type CoreTableInputOption,
+    type CoreQuery,
+    type CoreMatrix,
+    OwidTableSlugs,
+    type EntityName,
+    type EntityCode,
+    type EntityId,
+    type OwidColumnDef,
+    OwidEntityNameColumnDef,
+    OwidEntityIdColumnDef,
+    OwidEntityCodeColumnDef,
+    StandardOwidColumnDefs,
+    type OwidRow,
+    type OwidVariableRow,
+    ColumnTypeNames,
+    type ColumnColorScale,
+    type CoreColumnDef,
+    ErrorValue,
+} from "./domainTypes/CoreTableTypes.js"
